feat: extract Skeleton to a shared component
