+++ src/app/backups/components/BackupsAsFoldersList/BackupsAsFoldersList.tsx
@@ -6,8 +6,8 @@ import { DriveItemData } from '../../../drive/types';
 import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
 import Empty from '../../../shared/components/Empty/Empty';
 import List from '../../../shared/components/List';
-import { skinSkeleton } from 'app/drive/components/DriveExplorer/DriveExplorerList/DriveExplorerList';
 import { ListItemMenu } from 'app/shared/components/List/ListItem';
+import { skinSkeleton } from 'app/shared/Skeleton';
 
 export default function BackupsAsFoldersList({
   className = '',
