+++ src/app/drive/components/DriveExplorer/DriveExplorerList/DriveExplorerList.tsx
@@ -35,6 +35,7 @@ import {
   contextMenuWorkspaceFile,
   contextMenuWorkspaceFolder,
 } from './DriveItemContextMenu';
+import { skinSkeleton } from 'app/shared/Skeleton';
 
 interface DriveExplorerListProps {
   folderId: string;
@@ -546,13 +547,3 @@ export default connect((state: RootState) => ({
     state.ui.isItemDetailsDialogOpen ||
     state.ui.isPreferencesDialogOpen,
 }))(DriveExplorerList);
-
-export const skinSkeleton = [
-  <div className="mr-3 flex w-full flex-row items-center space-x-4">
-    <div className="h-8 w-8 rounded-md bg-gray-5" />
-    <div className="h-4 w-full rounded bg-gray-5" />
-  </div>,
-  <div className="h-4 w-64 rounded bg-gray-5" />,
-  <div className="ml-3 h-4 w-24 rounded bg-gray-5" />,
-  <div className="ml-4 h-4 w-20 rounded bg-gray-5" />,
-];
