+++ src/app/shared/Skeleton/index.tsx
@@ -1,0 +1,9 @@
+export const skinSkeleton = [
+  <div className="mr-3 flex w-full flex-row items-center space-x-4">
+    <div className="h-8 w-8 rounded-md bg-gray-5" />
+    <div className="h-4 w-full rounded bg-gray-5" />
+  </div>,
+  <div className="h-4 w-64 rounded bg-gray-5" />,
+  <div className="ml-3 h-4 w-24 rounded bg-gray-5" />,
+  <div className="ml-4 h-4 w-20 rounded bg-gray-5" />,
+];
